MDEV-34705: Binlog in Engine: Refactor rpl_gtid.h
Move rpl_gtid and rpl_binlog_state_base into separate rpl_gtid_base.h include
that can be used from engines implementing the binlog interface.
